Replace hash (#) character by underscore (_).
I am using bugwarrior to pull issues from redmine, and
bugwarrior uses the #-character in task descriptions, like e.g:

(bw)Is#5543 - Personen die stoppen en terug begin .. https://websites.chiro.be/issues/5543

If this is synced as-is to hamster, an activity '(bw)Is' is started,
and that is not what I want.

By replacing the # by an _, I work around the problem.